Finish KdTree implementation.
I wonder if this actually works.
It's probably a little to defensive though and might use one unnecessary call to a shape's intersect function. On the other hand, it is guaranteed to split the scene into boxes containing one single shape per box.
I don't know if this will give greater performance than intersecting with a series of shapes stored in the same leaf.
My guess is that the hit function for bounding boxes may be faster than hitting several shapes, but we can only tell once a test has been put in place.